Move listing utils to KubeResourceSeeker
