+++ mirrord/kube/src/api/kubernetes.rs
@@ -30,7 +30,6 @@ use crate::{
     error::{KubeApiError, Result},
 };
 
-pub mod list;
 #[cfg(feature = "portforward")]
 pub mod portforwarder;
 pub mod rollout;
